Renamed `recordingchannelgroup` module to `channelindex`
