+++ neo/core/channelindex.py
